Remove final support for dev mode, unify new mocks
