+++ .github/workflows/ci-tests.yml
@@ -17,7 +17,7 @@ env:
   PAGINATE: 30
   CACHETIME: 600000
   GCLOUD_STORAGE_BUCKET: ""
-  GOOGLE_APPLICATION_CREDENTIALS: ""
+  GOOGLE_APPLICATION_CREDENTIALS: "nofile"
   GH_CLIENTID: ""
   GH_CLIENTSECRET: ""
   GH_USERAGENT: "Pulsar-Edit Tester Bot"
